+++ R/corrOrthologTE.R
@@ -5,13 +5,16 @@
 #' just returning them, please specify the fileDir and fileName with the 
 #' extension .csv. The default fileName is TEKRABber_geneTECorrReusult.csv.
 #' @usage corrOrthologTE(geneInput, teInput, corrMethod = "pearson", 
-#' padjMethod = "fdr", fileDir=NULL, fileName="TEKRABber_geneTECorrResult.csv")
+#' padjMethod = "fdr", numCore=1, fileDir=NULL, 
+#' fileName="TEKRABber_geneTECorrResult.csv")
 #' @param geneInput gene count input for correlation from using DECorrInputs()
 #' @param teInput te count input for correlation from using DECorrInputs()
 #' @param corrMethod correlation method, including pearson, kendall, spearman. 
 #' Default is pearson.
 #' @param padjMethod method to return adjusted p-value, and default is fdr. 
 #' See ?p.adjust
+#' @param numCore number of cores to run parallel. Default is 1. You can use 
+#' detectCores() to get how many cores you can use.
 #' @param fileDir the name of directory for saving output files. 
 #' Default is NULL.
 #' @param fileName the name for saving output files. 
@@ -20,6 +23,8 @@
 #' @useDynLib TEKRABber
 #' @importFrom stats p.adjust
 #' @importFrom Rcpp sourceCpp
+#' @importFrom doParallel registerDoParallel stopImplicitCluster
+#' @import foreach
 #' @export
 #' 
 #' @examples
@@ -44,6 +49,7 @@
 #' controlCorr <- corrOrthologTE(
 #'     geneInput = resultDE$geneCorrInputRef[c(1:10),],
 #'     teInput = resultDE$teCorrInputRef[c(1:10),],
+#'     numCore = 1,
 #'     corrMethod = "pearson",
 #'     padjMethod = "fdr"
 #' )
@@ -52,15 +58,37 @@ corrOrthologTE <- function(
     geneInput, 
     teInput, 
     corrMethod = "pearson", 
-    padjMethod="fdr", 
+    padjMethod="fdr",
+    numCore=1,
     fileDir=NULL, 
     fileName="TEKRABber_geneTECorrResult.csv"){
-
-    df.ortholog <- t(geneInput)
-    df.te <- t(teInput)
     
-    df.corr <- rcpp_corr(df.ortholog, df.te, corrMethod)
-    colnames(df.corr)[c(1,2)] <- c("geneName", "teName")
+    # register backend
+    registerDoParallel(numCore)
+    
+    # create empty dataframe
+    df.corr <- data.frame(matrix(ncol=4, nrow=0))
+    colnames(df.corr) <- c("geneName", "teName", "coef", "pvalue")
+    
+    # calculate correlation with parallel
+    df.corr <- foreach(i=1:nrow(geneInput), .combine=rbind) %dopar% {
+        foreach(j=1:nrow(teInput), .combine=rbind) %dopar% {
+            num1 <- as.numeric(geneInput[i,])
+            num2 <- as.numeric(teInput[j,])
+            corr <- cor.test(num1, num2, method=corrMethod)
+            rbind(
+                data.frame(geneName=rownames(geneInput)[i], 
+                           teName=rownames(teInput)[j], 
+                           coef=corr$estimate, 
+                           pvalue=corr$p.value),
+                df.corr
+            )
+        }
+    }
+    
+    stopImplicitCluster()
+    
+    # calculate p-adjusted value
     df.corr$padj <- p.adjust(
         df.corr$pvalue,
         method=padjMethod