+++ R/zzz.R
@@ -1,3 +1,9 @@
 .onLoad <- function(libname, pkgname) {
-    packageStartupMessage("Welcome to TEKRABber version 1.6.0")
+    messages <- c(
+        "Welcome to TEKRABber version 1.8.0",
+        "+ New function: prepareRMSK() for getting repeatmasker",
+        "+ New parameter: `numCore` in corrOrthologTE() for parallel computing"
+    )
+    
+    packageStartupMessage(paste(messages, collapse = "\n"))
 }